requirements.txt
Flask
python-docx
PyMuPDF
pytesseract
Pillow
nltk
transformers
torch
keybert
sentence-transformers
scikit-learn
werkzeug
